Modified logger stuff, added more docs
I got rid of the LOGGER static variable and most of the 'log'
convienence methods in JukeAlert, because they were creating/using their
own logger when we get a Logger instance from bukkit through the
JavaPlugin interface we implement. So I just left log(String message) to
use the bukkit built in logger. Same thing in JukeAlertListener, i got
rid of the getter function in JukeAlert and made the jaLogger public so
JukeAlertListener just accesses the jaLogger instance variable. Database
also now requires a Logger param for the constructor, which is the
JukeAlert logger for now